Move classpath fingerprinting to its own package
