+++ subprojects/core/src/main/java/org/gradle/internal/service/scopes/TaskExecutionServices.java
@@ -38,8 +38,8 @@ import org.gradle.internal.concurrent.ExecutorFactory;
 import org.gradle.internal.concurrent.ParallelismConfigurationManager;
 import org.gradle.internal.fingerprint.FingerprintCompareStrategy;
 import org.gradle.internal.fingerprint.HistoricalFileCollectionFingerprint;
+import org.gradle.internal.fingerprint.classpath.impl.ClasspathCompareStrategy;
 import org.gradle.internal.fingerprint.impl.AbsolutePathFingerprintCompareStrategy;
-import org.gradle.internal.fingerprint.impl.ClasspathCompareStrategy;
 import org.gradle.internal.fingerprint.impl.DefaultHistoricalFileCollectionFingerprint;
 import org.gradle.internal.fingerprint.impl.EmptyHistoricalFileCollectionFingerprint;
 import org.gradle.internal.fingerprint.impl.IgnoredPathCompareStrategy;
