+++ subprojects/platform-play/src/main/java/org/gradle/play/internal/toolchain/DefaultPlayToolProvider.java
@@ -17,7 +17,7 @@
 package org.gradle.play.internal.toolchain;
 
 import org.gradle.internal.file.PathToFileResolver;
-import org.gradle.internal.fingerprint.ClasspathFingerprinter;
+import org.gradle.internal.fingerprint.classpath.ClasspathFingerprinter;
 import org.gradle.language.base.internal.compile.CompileSpec;
 import org.gradle.language.base.internal.compile.Compiler;
 import org.gradle.play.internal.javascript.GoogleClosureCompiler;
