+++ subprojects/platform-play/src/test/groovy/org/gradle/play/internal/DefaultPlayToolChainTest.groovy
@@ -23,7 +23,7 @@ import org.gradle.api.artifacts.Dependency
 import org.gradle.api.artifacts.ResolveException
 import org.gradle.api.artifacts.dsl.DependencyHandler
 import org.gradle.internal.file.PathToFileResolver
-import org.gradle.internal.fingerprint.ClasspathFingerprinter
+import org.gradle.internal.fingerprint.classpath.ClasspathFingerprinter
 import org.gradle.internal.text.TreeFormatter
 import org.gradle.language.scala.ScalaPlatform
 import org.gradle.play.internal.toolchain.DefaultPlayToolChain
